Replace ConstCertContextHandle with CertContextHandle - they really were not different.
Enhance Handle class so it can deal with multi word handles (like SecHandle types) and redefine CredentialHandle and SecurityContextHandle in terms of Handle. This required the traits class to be able to specify Equals and Less functions so that multi word handles could be defined and used.

Some functions were redefined as methods, so, for example get(h) becomes h.get(), which makes intellisense more helpful.