Set a consistent set of options, in particular: Warning level 4 Treat warnings as errors, Draft (std:c++latest) Conformance mode: Yes(/permissive-) Enforce type conversion rules: Yes (/Zc:rvalueCast)
Added various casts to casts to make this compile, since much of the code calls ancient Win32 functions that are not const correct. There's a small risk this code might break in future (for example if if the compiler decides to handle string constants differently) but that's unlikely, since such a change would break exissting code.