Allow an Ansi build of StreamServer.
This still uses a Unicode build of SSLServer so to permit that SSLServer no longer uses MFC - it uses ATL instead, which leads to a few minor changes as method names or parameters are a little different. A fw Windows calls are now explicitly "W" (wide) versions.

There's now a "Debug Ansi" build which builds SimpleClient, SteamClient and StreamServer using multibyte (ansi) and SSLClient and SSLServer using Unicode.