Created a DLL for both client and server
In an application that can be both client and server, it wasn't possible to use both SSLClient and SSLserver DLL's, because the externals in the common source files would be in both DLL's and would cause a multiple defined symbol error when linking the project.

A new project that has all classes (Common, Client and Server) can now be used instead of either of the other main projects (SSLClient / SSLServer) or to have both client as well as server functionality.

I ran into a problem that might explain why this wasn't done before: The CredentialTraits::Close and SecurityContextTraits::Close used different code depending on what project they were part of. The code was basically the same: They called functions in g_SSPI; however both CSSLServer and CSSLClient had a static g_SSPI that was initialized once, as needed.

I moved the g_SSPI member and SSPI() function of CSSLClient and CSSLServer to a new CSSLCommon base class, and derived the CSSLClient and CSSLServer from the CSSLCommon class so everything still basically works as before.

NOTE: The new SChanSSL project uses most of the same files as the CSSLServer and CSSLClient projects (the framework.h, pch.c/h and targetver.h files are exceptions) so any changes in any projects will be reflected in all other projects. No files were copied from one place to another.

This is not fully tested yet but I'll get to that after this.