Removed remaining dependencies on MFC
* Replaced time-related members with classes from <chrono>
* Replaced uses of CEvent with EventWrapper; moved EventWrapper.h to the Common/Include folder and added SetEvent, ResetEvent and PulseEvent methods
* Replaced uses of CStringA with std::string
* Removed CCriticalSection and replaced the member it was protecting with std::atomic_int volatile

Also changed project options for all configurations so all output files and intermediate files end up in two trees Win32 and x64. Cleaning the tree to establish a clean recompile is as easy as deleting those two.

Also changed some project configurations so they can be built with VS2017 (toolset v141). I missed a few of them before.

Also renamed the SChanSSL project files to StreamSSL.* and changed the sample projects to depend on it, instead of on the Client and Server library projects

Also cleaned up the configuration manager which was a mess. For many platform/target combinations, the wrong projects were being built.