Improvements to Sample Code
The primary functional change is that the sample server side code no longer sends an initial greeting message. Instead it behaves more like a web server which does not send an initial message until it gets a request message from a client.

A consequence of this is that the SimpleClient application can be pointed at the server and it will just read the initial message and reject it with an "unexpected message" error then shut down the connection, much as a web server would if it did not understand the initial message.

The timing of the last message from StreamClient has been improved so there's no longer a race condition in the server that causes it to miss that message sometimes.

The rest is improvements to the progress reporting messages to do a better job of reporting what text was received.