Implement GetTlsVersion
Introduce `getTlsVersionFromProtocol` in `CSSLHelper` to convert protocol masks to two-digit TLS version numbers. Update `CSSLClient` and 'CSSLServer' to store and retrieve the TLS version, and modify `SSPINegotiate` in both classes to set this version based on the connection's protocol.

Enhance the various sample clients output to display the TLS version upon successful connection.